+++ examples/quick-start/main.go
@@ -35,6 +35,10 @@ func main() {
 		os.Exit(1)
 	}
 
-	detail := tasker.PostTask("Startup").Wait().GetDetail()
+	detail, err := tasker.PostTask("Startup").Wait().GetDetail()
+	if err != nil {
+		fmt.Println("Failed to get task detail:", err)
+		os.Exit(1)
+	}
 	fmt.Println(detail)
 }